refactor: keep artifact naming in config
