+++ pkg/config/artifacts_test.go
@@ -1,4 +1,4 @@
-package release
+package config
 
 import "testing"
